Interactive: Add new watchdog timeout feature
Interactive sessions can fail in a number of different ways, including:

1) Superuser request is denied, times out, or fails; "su" exits without
starting our shell.  Maybe the user mis-clicked, or maybe he just didn't
notice the prompt until it was too late.

2) The shell crashes / terminates / runs out of memory / etc, e.g. the
caller sends an "exit" command and so libsuperuser never sees its markers
on stdout/stderr.

3) A legitimate command blocks or hangs ("cat", "sleep 10000"), or an
incomplete command gets the shell stuck ("if").

We can't detect or prevent most of these, but at least we can make sure
the client gets a callback after N seconds if his command gets stuck.  We
can also guarantee that he will either see a timeout callback OR a
"command finished" callback, not both.

Note that watchdogTimeout is not declared as final, because we may want
to use a longer timeout for the initial command (the only command that
can generate an interactive SU prompt).

Signed-off-by: Jorrit Jongma <git@jongma.org>