Create new available() method for Interactive shell
Similar to SU.available(), check to see if the current interactive session
is responding to basic shell commands.  If it's a root session, also check
the reported UID.

An application could call this at startup time, and modify its behavior
(or abort) if the root shell could not be started.

Like most of the other Interactive calls, this is non-blocking.