Garbage collection protection
As actual shell close may occur in the background later after close()
calls have been issued, it might be possible the last reference to a
Shell.Threaded instance is removed before cleanup is finished, or even
during the cleanup.

This commit should prevent that from happening, but as the problem
cannot be reliably reproduced, the jury is still out on this fix.

@see 81